Time: 42 ms (5.12%), Space: 5.8 MB (76.44%) - LeetHub

+++ 2240-number-of-ways-to-buy-pens-and-pencils/2240-number-of-ways-to-buy-pens-and-pencils.cpp
@@ -6,7 +6,7 @@ public:
             val = total-cost1*i;
             if(val<0)
                 break;
-            ans+= (long long)floor((long)val/(long)cost2)+1;
+            ans+= (long)floor((long)val/(long)cost2)+1;
         }
         return ans;
     }
